+++ utils/misc_test.go
@@ -19,3 +19,9 @@ func TestMinInt(t *testing.T) {
 	assert.Equal(t, 0, utils.MinInt(1, 0))
 	assert.Equal(t, -1, utils.MinInt(1, -1))
 }
+
+func TestDeriveCountryFromTel(t *testing.T) {
+	assert.Equal(t, "RW", utils.DeriveCountryFromTel("+250788383383"))
+	assert.Equal(t, "EC", utils.DeriveCountryFromTel("+593979000000"))
+	assert.Equal(t, "", utils.DeriveCountryFromTel("1234"))
+}