+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+v0.14.7
+----------
+ * Replicate channel matching logic from RP
+
+v0.14.6
+----------
+ * Don't strip URN params when parsing
+
+v0.14.5
+----------
+ * Migrate legacy webhook calls to POST if they don't have a method set
+
+v0.14.4
+----------
+ * Simplify string tokenization and add more tests
+
 v0.14.3
 ----------
  * Fix message attachment expressions not being URL encoded